+++ src/tools/shader.ts
@@ -3,11 +3,11 @@ import { defaultUniforms } from "../defaults";
 
 export default class Shader extends Compilable {
   constructor(source: string) {
+    super(source);
+
     // Add default uniforms to requirements
     defaultUniforms.forEach((type, name) => {
       this.requirements.uniforms.set(name, { name, type, line: 0 });
     });
-
-    super(source);
   }
 }
